Move pipeline classes into separate files
