+++ ceci/pipeline/__init__.py
@@ -1,0 +1,10 @@
+"""Module with core pipeline functionality """
+
+from .cwl import CWLPipeline
+from .pipeline import Pipeline, override_config
+from .parsl import ParslPipeline
+from .dry_run import DryRunPipeline
+from .flow_chart import FlowChartPipeline
+from .mini import MiniPipeline
+from .file_manager import FileManager
+from .sec import StageExecutionConfig
